third_party/blink/web_tests/platform/generic/fast/forms/number/number-spinbutton-gets-disabled-or-readonly-expected.txt

 

This tests if the value of a number input form does not continue to increase/decrease in the following scenario.
(1) Click the spin button of the input form.
(2) Hook the 'mouseup' event and disable the input form.
(3) Enable the input form after some delay.
To run this test manually, input any value in the input form and then click the spin-down button. At this point, click the spin-down button quickly and do not move the cursor from the spin-down button after the click. If the value decreases by just 1, this test passes.



Test on a readonly number input form:
delay = 1 ms
PASS input.value is "1234566"
delay = 10 ms
PASS input.value is "1234566"
delay = 100 ms
PASS input.value is "1234566"


Test on a disabled number input form:
delay = 1 ms
PASS input.value is "1234566"
delay = 10 ms
PASS input.value is "1234566"
delay = 100 ms
PASS input.value is "1234566"
PASS successfullyParsed is true

TEST COMPLETE
